Segregate testing as well
