+++ mslib/utils/_tests/test_config.py
@@ -1,0 +1,172 @@
+# -*- coding: utf-8 -*-
+"""
+
+    mslib.utils._tests.test_config
+    ~~~~~~~~~~~~~~~~~~~~~~~
+
+    This module provides pytest functions to tests mslib.utils.config
+
+    This file is part of mss.
+
+    :copyright: Copyright 2016-2017 Reimar Bauer
+    :copyright: Copyright 2016-2021 by the mss team, see AUTHORS.
+    :license: APACHE-2.0, see LICENSE for details.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+"""
+import logging
+import pytest
+import os
+import fs
+import mslib.utils.config as config
+import mslib.utils as utils
+from mslib._tests.constants import MSS_CONFIG_PATH
+from mslib._tests.utils import create_mss_settings_file
+from mslib._tests import constants
+
+LOGGER = logging.getLogger(__name__)
+
+
+class TestSettingsSave(object):
+    """
+    tests save_settings_qsettings and load_settings_qsettings from ./utils.py
+    # TODO make sure do a clean setup, not inside the 'mss' config file.
+    """
+    tag = "test_automated"
+
+    def test_save_settings(self):
+        settings = {'foo': 'bar'}
+        config.save_settings_qsettings(self.tag, settings, ignore_test=True)
+
+    def test_load_settings(self):
+        settings = config.load_settings_qsettings(self.tag, ignore_test=True)
+        assert isinstance(settings, dict)
+        assert settings["foo"] == "bar"
+
+
+class TestConfigLoader(object):
+    """
+    tests config file for client
+    """
+
+    def teardown(self):
+        if fs.open_fs(MSS_CONFIG_PATH).exists("mss_settings.json"):
+            fs.open_fs(MSS_CONFIG_PATH).remove("mss_settings.json")
+
+    def test_default_config(self):
+        data = config.config_loader()
+        assert isinstance(data, dict)
+        assert data["num_labels"] == 10
+        assert data["num_interpolation_points"] == 201
+
+    def test_default_config_dataset(self):
+        data = config.config_loader(dataset="num_labels")
+        assert data == 10
+        # defined value and not a default one
+        data = config.config_loader(dataset="num_labels")
+        assert data == 10
+
+    def test_default_config_wrong_file(self):
+        # return default if no access to config file given
+        with pytest.raises(utils.FatalUserError):
+            config.config_loader(config_file="foo.json")
+
+    def test_sample_config_file(self):
+        utils_path = os.path.dirname(os.path.abspath(utils.__file__))
+        config_file = os.path.join(utils_path, '..', '..', 'docs', 'samples', 'config', 'mss',
+                                   'mss_settings.json.sample')
+        data = config.config_loader(config_file=config_file, dataset="new_flighttrack_flightlevel")
+        assert data == 250
+        with pytest.raises(KeyError):
+            config.config_loader(config_file=config_file, dataset="UNDEFINED")
+        with pytest.raises(KeyError):
+            assert config.config_loader(config_file=config_file, dataset="UNDEFINED")
+        with pytest.raises(utils.FatalUserError):
+            config_file = os.path.join(utils_path, '..', '..', 'docs', 'samples', 'config', 'mss',
+                                       'not_existing_mss_settings.json.sample')
+            config.config_loader(config_file=config_file)
+
+    def test_config_file_cached(self, caplog):
+        with caplog.at_level(logging.INFO):
+            config.config_loader()
+        assert 'Default MSS configuration in place, no user settings, see http://mss.rtfd.io/en/stable/usage.html' \
+               in caplog.text
+        assert constants.CACHED_CONFIG_FILE is None
+
+    def test_existing_empty_config_file(self):
+        """
+        on a user defined empty mss_settings_json this test should return the default value for num_labels
+        """
+        create_mss_settings_file('{ }')
+        if not fs.open_fs(MSS_CONFIG_PATH).exists("mss_settings.json"):
+            pytest.skip('undefined test mss_settings.json')
+        with fs.open_fs(MSS_CONFIG_PATH) as file_dir:
+            file_content = file_dir.readtext("mss_settings.json")
+        assert ":" not in file_content
+        config_file = fs.path.combine(MSS_CONFIG_PATH, "mss_settings.json")
+        data = config.config_loader(config_file=config_file)
+        assert data["num_labels"] == 10
+        num_labels = config.config_loader(config_file=config_file, dataset="num_labels")
+        assert num_labels == 10
+        # this overwrites the builtin default value
+        num_labels = config.config_loader(config_file=config_file, dataset="num_labels")
+        assert num_labels == 10
+        with pytest.raises(KeyError):
+            config.config_loader(config_file=config_file, dataset="UNDEFINED")
+        with pytest.raises(KeyError):
+            assert config.config_loader(config_file=config_file, dataset="UNDEFINED")
+
+    def test_existing_config_file_different_parameters(self):
+        """
+        on a user defined mss_settings_json without a defined num_labels this test should return its default value
+        """
+        create_mss_settings_file('{"num_interpolation_points": 20 }')
+        if not fs.open_fs(MSS_CONFIG_PATH).exists("mss_settings.json"):
+            pytest.skip('undefined test mss_settings.json')
+        with fs.open_fs(MSS_CONFIG_PATH) as file_dir:
+            file_content = file_dir.readtext("mss_settings.json")
+        assert "num_labels" not in file_content
+        config_file = fs.path.combine(MSS_CONFIG_PATH, "mss_settings.json")
+        data = config.config_loader(config_file=config_file)
+        assert data["num_labels"] == 10
+        num_labels = config.config_loader(config_file=config_file, dataset="num_labels")
+        assert num_labels == 10
+        num_interpolation_points = config.config_loader(config_file=config_file, dataset="num_interpolation_points")
+        assert num_interpolation_points == 20
+        data = config.config_loader(config_file=config_file)
+        assert data["num_interpolation_points"] == 20
+        with pytest.raises(KeyError):
+            config.config_loader(config_file=config_file, dataset="UNDEFINED")
+        with pytest.raises(KeyError):
+            assert config.config_loader(config_file=config_file, dataset="UNDEFINED")
+
+    def test_existing_config_file_defined_parameters(self):
+        """
+        on a user defined mss_settings_json without a defined num_labels this test should return its default value
+        """
+        create_mss_settings_file('{"num_interpolation_points": 201, "num_labels": 10 }')
+        if not fs.open_fs(MSS_CONFIG_PATH).exists("mss_settings.json"):
+            pytest.skip('undefined test mss_settings.json')
+        with fs.open_fs(MSS_CONFIG_PATH) as file_dir:
+            file_content = file_dir.readtext("mss_settings.json")
+        assert "num_labels" in file_content
+        config_file = fs.path.combine(MSS_CONFIG_PATH, "mss_settings.json")
+        num_labels = config.config_loader(config_file=config_file, dataset="num_labels")
+        assert num_labels == 10
+        # this overwrites the given value
+        num_labels = config.config_loader(config_file=config_file, dataset="num_labels")
+        assert num_labels == 10
+        with pytest.raises(KeyError):
+            config.config_loader(config_file=config_file, dataset="UNDEFINED")
+        with pytest.raises(KeyError):
+            assert config.config_loader(config_file=config_file, dataset="UNDEFINED")
